HSEARCH-4876 Move cycle detection tests to its own class
